Abstract formatDateTimeRelative into datetime utils
